--- presentaiton/src/main/java/com/goms/presentaiton/GOMSApplication.kt
@@ -1,8 +1,0 @@
-package com.goms.presentaiton
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class GOMSApplication: Application() {
-}
