--- presentaiton/src/main/java/com/goms/presentaiton/utils/extension/LifeCycleOwnerExtension.kt
@@ -1,14 +1,0 @@
-package com.goms.presentaiton.utils.extension
-
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.LifecycleOwner
-import androidx.lifecycle.lifecycleScope
-import androidx.lifecycle.repeatOnLifecycle
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.launch
-
-fun LifecycleOwner.repeatOnStart(block: suspend CoroutineScope.()-> Unit){
-    lifecycleScope.launch{
-        lifecycle.repeatOnLifecycle(Lifecycle.State.STARTED, block)
-    }
-}
