+++ presentation/build.gradle.kts
@@ -51,6 +51,8 @@ android {
 }
 
 dependencies {
+    implementation(project(":domain"))
+    implementation(project(":di"))
 
     implementation("androidx.core:core-ktx:1.9.0")
     implementation("androidx.lifecycle:lifecycle-runtime-ktx:2.6.1")
