+++ presentation/src/main/java/com/example/goms_android_v2_admin/utils/ErrorHandling.kt
@@ -1,4 +1,4 @@
-package com.goms.presentaiton.utils
+package com.example.goms_android_v2_admin.utils
 
 
 import android.util.Log
