+++ presentation/src/main/res/mipmap-xxxhdpi/ic_launcher.webp
