+++ settings.gradle.kts
@@ -14,8 +14,7 @@ dependencyResolutionManagement {
 }
 
 rootProject.name = "GOMS-Android-V2-Admin"
-include(":app")
+include(":presentation")
 include(":data")
 include(":di")
 include(":domain")
-include(":presentaiton")
